+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.36.0] - 2026-02-07
+
 ### Added
 - **kiarina-lib-atlassian**: Initial release with Atlassian API configuration management